Split common server (#481)
