+++ lib/src/server_cache.dart
@@ -232,4 +232,3 @@ class InMemoryCache implements ServerCache {
   @override
   Future<void> shutdown() => Future<void>.value();
 }
-
