+++ test/bench_test.dart
@@ -27,8 +27,7 @@ void defineTests() {
     });
 
     test('many', () {
-      final harness =
-          BenchmarkHarness(asJson: true, logger: (_) => null);
+      final harness = BenchmarkHarness(asJson: true, logger: (_) => null);
       final benchmarks = [MockBenchmark(), MockBenchmark()];
 
       return harness.benchmark(benchmarks).then((_) {
